@@ -1,6 +1,6 @@
 module github.com/opcr-io/policy
 
-go 1.17
+go 1.19
 
 // replace github.com/aserto-dev/go-utils ../go-utils
 
