@@ -9,7 +9,7 @@ type InspectCmd struct {
 func (c *InspectCmd) Run(g *Globals) error {
 	err := g.App.Inspect(c.Policy)
 	if err != nil {
-		return errors.Wrap(err, "Failed to inspect policy.")
+		return errors.Wrap(err, "failed to inspect policy")
 	}
 
 	<-g.App.Context.Done()
