@@ -1,8 +1,6 @@
 package main
 
-import (
-	"errors"
-)
+import "github.com/pkg/errors"
 
 type PullCmd struct {
 	Policies []string `arg:"" name:"policy" help:"Policies to pull from the remote registry."`
@@ -21,7 +19,7 @@ func (c *PullCmd) Run(g *Globals) error {
 	<-g.App.Context.Done()
 
 	if errs != nil {
-		return errors.New("failed to pull one or more policies")
+		return errors.Wrap(errs, "failed to pull one or more policies")
 	}
 
 	return nil
