@@ -10,7 +10,7 @@ type ReplCmd struct {
 func (c *ReplCmd) Run(g *Globals) error {
 	err := g.App.Repl(c.Policy, c.MaxErrors)
 	if err != nil {
-		return errors.Wrap(err, "There was an error running the OPA runtime.")
+		return errors.Wrap(err, "there was an error running the OPA runtime")
 	}
 
 	<-g.App.Context.Done()
