@@ -5,7 +5,7 @@ import "github.com/opcr-io/policy/pkg/version"
 type VersionCmd struct {
 }
 
-func (c *VersionCmd) Run(g *Globals) error {
+func (c *VersionCmd) Run(g *Globals) error { //nolint:unparam
 	v := version.GetInfo()
 
 	g.App.UI.Normal().
