@@ -11,7 +11,7 @@ import (
 )
 
 // CC contains dependencies that are cross cutting and are needed in most
-// of the providers that make up this application
+// of the providers that make up this application.
 type CC struct {
 	Context    context.Context
 	Config     *config.Config
@@ -27,7 +27,7 @@ var (
 	singletonErr error
 )
 
-// NewCC creates a singleton CC
+// NewCC creates a singleton CC.
 func NewCC(logOutput logger.Writer, errOutput logger.ErrWriter, configPath config.Path, overrides config.Overrider) (*CC, func(), error) {
 	once.Do(func() {
 		cc, cleanup, singletonErr = buildCC(logOutput, errOutput, configPath, overrides)
@@ -40,7 +40,7 @@ func NewCC(logOutput logger.Writer, errOutput logger.ErrWriter, configPath confi
 }
 
 // NewTestCC creates a singleton CC to be used for testing.
-// It uses a fake context (context.Background)
+// It uses a fake context (context.Background).
 func NewTestCC(logOutput logger.Writer, errOutput logger.ErrWriter, configPath config.Path, overrides config.Overrider) (*CC, func(), error) {
 	once.Do(func() {
 		cc, cleanup, singletonErr = buildTestCC(logOutput, errOutput, configPath, overrides)
