@@ -7,14 +7,14 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/manager/signals"
 )
 
-// ErrGroupAndContext wraps a context and an error group
+// ErrGroupAndContext wraps a context and an error group.
 type ErrGroupAndContext struct {
 	Ctx      context.Context
 	ErrGroup *errgroup.Group
 	Cancel   context.CancelFunc
 }
 
-// NewContext creates a context that responds to user signals
+// NewContext creates a context that responds to user signals.
 func NewContext() *ErrGroupAndContext {
 	errGroup, ctxErr := errgroup.WithContext(signals.SetupSignalHandler())
 	ctx, cancelFunc := context.WithCancel(ctxErr)
@@ -26,7 +26,7 @@ func NewContext() *ErrGroupAndContext {
 	}
 }
 
-// NewTestContext creates a context that can be used for testing
+// NewTestContext creates a context that can be used for testing.
 func NewTestContext() *ErrGroupAndContext {
 	errGroup, ctxErr := errgroup.WithContext(context.Background())
 	ctx, cancelFunc := context.WithCancel(ctxErr)
