@@ -6,7 +6,7 @@ import (
 	"time"
 )
 
-// values set by linker using ldflag -X
+// values set by linker using ldflag -X.
 var (
 	ver    string // nolint:gochecknoglobals // set by linker
 	date   string // nolint:gochecknoglobals // set by linker
